+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - **`[knf/validators/time]`** Added timezone validator
 - **`[knf/validators]`** Code refactoring
+- **`[timeutil]`** Code refactoring
 
 ### [13.20.1](https://kaos.sh/ek/13.20.1)